+++ core/exports/effects.ts
@@ -8,7 +8,7 @@ export {
   runWith,
   transformerFor,
 } from "../src/effects/effects.ts";
-export { io, pluginIO } from "../src/effects/io.ts";
+export { io } from "../src/effects/io.ts";
 export { config, denoConfig } from "../src/effects/config.ts";
 export { importmap } from "../src/effects/impormap.ts";
 export { hot } from "../src/effects/hot-update.ts";
